+++ docs/Changelog.md
@@ -21,6 +21,8 @@ Nussknacker versions
 * [#704](https://github.com/TouK/nussknacker/pull/704) Preloaded creator panel node icons
 * [#943](https://github.com/TouK/nussknacker/pull/943) Literal min / max validators
 * [#973](https://github.com/TouK/nussknacker/pull/973) Textarea editor
+* [#981](https://github.com/TouK/nussknacker/pull/981) Refactor Kafka and KafkaAvro serialization package
+
 
 0.1.2
 ------------------------